Calendar: Add a function to get a LeiningDate from an arbitrary parsha name
This would be used to support existing URLs (eg, for bookmarks or links from other sites) that only specify a parsha name.  It is unclear how important that is.

However, this code is currently broken (the unit test fails) because the underlying hebcal API returns null if you pass a single parsha that happens to be double that year.  IOW, the test fails on (eg) `Matot` because the year in the test contains `Matot-Masei`.
See hebcal/hebcal-es6#467